Use an advice instead of defining new commands; delete our keymap
This is the same ideas as in commit ae93720, except more generalised.
Instead of defining our own commands, we advise Dired to display a
preview after one of its standard commands is called.  This way we
avoid the problem where the key remap mechanism does not actually
remap all keys associated with a given command.[1][2]

[1] Thanks to Peter Prevos for reporting this in issue 1 on the GitHub
    mirror: <https://github.com/protesilaos/dired-preview/issues/1>.

[2] Thanks to Christian Tietze and Ed Hamilton for discussing this
    with me on the mailing list:
    <https://lists.sr.ht/~protesilaos/general-issues/%3Cm1zg4noej2.fsf%40christiantietze.de%3E>.
    Commit ae93720 by Christian Tietze is based on this discussion.