Fix how delayed-mode-hooks are handled
Before, the 'dired-preview--run-mode-hooks' would only be installed if
the preview was shown in a window.  This, however, led to a situation
where the buffer would be prepared for preview but never have time to
be displayed because the user visited the file in the meantime.  The
delayed hooks would never be called in that case.