Make dired-preview-display-file delete windows eagerly
The previous behaviour would produce more than one preview windows in
the case where (i) the Emacs frame displayed a preview per the
function 'dired-preview-display-action-alist-dwim', (ii) the frame was
resized, and (iii) a new preview was requested.

Thanks to Bruno Boal for bringing this matter to my attention and for
checking the workaround with me: <https://git.sr.ht/~bboal>.  This was
done in a private channel and the information is shared with
permission.