Stop adding 'no-other-window' parameter
The 'no-other-window' creates problems when the user selects the
window using the mouse, because Emacs disregards that setting.  As
such, we have inconsistent behaviour when we try to focus a preview
with the mouse versus the keyboard.

Thanks to Bruno Boal for bringing this matter to my attention:
<https://git.sr.ht/~bboal>.  This was done in a private channel and
the information is shared with permission.