Stop using a side window for previews
We want the user to be able to focus the window, either with the
keyboard or the mouse.  The window must then be a regular window.  Any
special restrictions feel counter-intuitive as the user will not know
that we had decided to set them in the first place.

Thanks to Bruno Boal for bringing this matter to my attention:
<https://git.sr.ht/~bboal>.  This was done in a private channel and
the information is shared with permission.