Refactor how parameters are set
This should fix inconsistencies with how preview windows are treated.
I want them to be deleted at all times as soon as we are out of Dired.

Before, this was not working with mouse clicking to focus a window, so
I tried to generally not close previews.  That caused effects that I
would consider counter-intuitive.  I prefer preview windows to be
"ephemeral" and thus be treated specially.