Implement garbage collection for buffer killing
This way we avoid killing buffers too eagerly.  The accumulated
buffers are still useful for bringing back up a preview shortly after
it has been removed from the preview window.

Thanks to Bruno Boal[1] for testing this with me.  It was done in a
private channel and the information is shared with permission.

[1] <https://git.sr.ht/~bboal>