Remove asterisk from if-let, when-let
Those are apparently not documented in the Emacs Lisp Reference Manual
and should thus not be considered analogous to 'let*'.  The 'if-let'
and 'when-let' are designed to be like 'let*'.  Not sure what the
others should be in this case...

Thanks to Philip Kaludercic for pointing this out on the emacs-devel
mailing list:
<https://lists.gnu.org/archive/html/emacs-devel/2023-07/msg00406.html>.