dired-preview: Add support for large file preview
* dired-preview.el (dired-preview--kill-large-buffers,
dired-preview--preview-p, dired-preview--add-to-previews,
dired-preview--chunk-size, dired-preview--find-large-file,
dired-preview--add-truncation-message,
dired-preview--large-files-alist, dired-preview--dispatch-file,
dired-preview--display-buffer): Preview files larger than
`dired-preview-max-size' by reading in only a leading chunk of
data. The size of this chunk is specified by
`dired-preview--chunk-size'.

The function `dired-preview--dispatch-file' is used to determine
how a file should be previewed.  Currently it distinguishes
between large files and not-large files, but this will be expanded
in the future to handle media previews, etc.

The function `dired-preview--find-large-file' is responsible for
setting up the preview.  The corresponding buffer is in
fundamental mode, set to read-only mode with undo disabled.  The
function `dired-preview--add-truncation-message' inserts a message
at the end of the previewed chunk.

In this process,
- the large file cannot be marked as being visited,
- `delayed-mode-hooks' is nil, and
- the buffer is modified,
thus we cannot use the criteria used for regular previews to find
buffers showing previews of large files.  We instead maintain a
separate list, `dired-preview--large-files-alist', to track them.