dired-preview: View binary files in hexl-mode
* dired-preview.el (dired-preview-hexl-toggle, dired-preview--find-large-file,
dired-preview-binary-as-hexl): Large files recognized by Emacs as
non-text (binary) are previewed using `hexl-mode'. This is generally preferable
to viewing the raw byte stream in a buffer.

The variable `dired-preview-binary-as-hexl' controls the default
behavior.

At any time, the user can toggle between the raw/text view and
hexl-mode view of a preview buffer by invoking
`dired-preview-hexl-toggle', which is bound to `C-c C-c' in
`dired-preview-mode-map'.